Create countTypeCICIDS2017.py
This is the attack type counter for dataset CICIDS 2017 as part of the trial process. This code will be utilized for other individual datasets. Keep in note that when accessing CICIDS 2017 there were multiple CSV files that needed to be combined --- To find that code refer to combinedCICIDS2017.py also found on this repository